# docs/openapi.json
{
  "components": {
    "schemas": {
      "AdminOverviewResponse": {
        "properties": {
          "active_forms": {
            "title": "Active Forms",
            "type": "integer"
          },
          "answers": {
            "title": "Answers",
            "type": "integer"
          },
          "failed_enqueue_reminders": {
            "title": "Failed Enqueue Reminders",
            "type": "integer"
          },
          "forms": {
            "title": "Forms",
            "type": "integer"
          },
          "reminders_by_status": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "Reminders By Status",
            "type": "object"
          },
          "stale_pending_reminders": {
            "title": "Stale Pending Reminders",
            "type": "integer"
          },
          "users": {
            "title": "Users",
            "type": "integer"
          }
        },
        "required": [
          "users",
          "forms",
          "active_forms",
          "answers",
          "reminders_by_status",
          "failed_enqueue_reminders",
          "stale_pending_reminders"
        ],
        "title": "AdminOverviewResponse",
        "type": "object"
      },
      "AdminReminderResponse": {
        "properties": {
          "delivery_count": {
            "title": "Delivery Count",
            "type": "integer"
          },
          "enqueue_status": {
            "title": "Enqueue Status",
            "type": "string"
          },
          "form_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Form Id"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "last_enqueue_error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Enqueue Error"
          },
          "next_run_at": {
            "format": "date-time",
            "title": "Next Run At",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "user_id": {
            "title": "User Id",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "user_id",
          "form_id",
          "title",
          "status",
          "enqueue_status",
          "delivery_count",
          "next_run_at",
          "updated_at"
        ],
        "title": "AdminReminderResponse",
        "type": "object"
      },
      "AdminUserResponse": {
        "properties": {
          "created_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "email": {
            "title": "Email",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "is_email_confirmed": {
            "title": "Is Email Confirmed",
            "type": "boolean"
          },
          "roles": {
            "items": {
              "type": "string"
            },
            "title": "Roles",
            "type": "array"
          },
          "timezone": {
            "title": "Timezone",
            "type": "string"
          }
        },
        "required": [
          "id",
          "email",
          "roles",
          "is_email_confirmed",
          "timezone"
        ],
        "title": "AdminUserResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/auth/login"
          }
        },
        "type": "oauth2"
      }
    }
  },
  "info": {
    "description": "FastAPI with PostgreSQL database.",
    "title": "ping me Project API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/admin/overview": {
      "get": {
        "operationId": "get_overview_admin_overview_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminOverviewResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Get Overview",
        "tags": [
          "Admin"
        ]
      }
    },
    "/admin/reminders/failed": {
      "get": {
        "operationId": "list_failed_reminders_admin_reminders_failed_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AdminReminderResponse"
                  },
                  "title": "Response List Failed Reminders Admin Reminders Failed Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "List Failed Reminders",
        "tags": [
          "Admin"
        ]
      }
    },
    "/admin/users": {
      "get": {
        "operationId": "list_users_admin_users_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AdminUserResponse"
                  },
                  "title": "Response List Users Admin Users Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "List Users",
        "tags": [
          "Admin"
        ]
      }
    },
...
  }
}